backend/requirements.txt
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7
Pillow==10.2.0 